Styling message input
