--- code/src/NewThought.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const NewThought = () => {
-  return (
-    <p className="new-message">
-      New message
-    </p>
-  )
-}
-
-export default NewThought;
