+++ code/src/components/thought/Message.js
