+++ code/src/components/thought/Thought.js
