+++ code/src/components/thought/ThoughtInfo.js
