+++ code/src/components/thought/Time.js
